move deployment and service to package resources
