+++ components/buildless-serverless/internal/controller/state/handle_deployment.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 	serverlessv1alpha2 "github.com/kyma-project/serverless/api/v1alpha2"
-	"github.com/kyma-project/serverless/internal/controller/deployment"
 	"github.com/kyma-project/serverless/internal/controller/fsm"
+	"github.com/kyma-project/serverless/internal/controller/resources"
 	appsv1 "k8s.io/api/apps/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -21,7 +21,7 @@ import (
 // - gitSources - stateFnGitCheckSources
 
 func sFnHandleDeployment(ctx context.Context, m *fsm.StateMachine) (fsm.StateFn, *ctrl.Result, error) {
-	m.State.Deployment = deployment.New(&m.State.Function, &m.FunctionConfig)
+	m.State.Deployment = resources.NewDeployment(&m.State.Function, &m.FunctionConfig)
 	builtDeployment := m.State.Deployment.Deployment
 
 	clusterDeployment, resultGet, errGet := getOrCreateDeployment(ctx, m, builtDeployment)
